+++ packages/nodejs/test/testHelpers.ts
@@ -4,6 +4,8 @@ import ConcordiumNodeClient from '../src/client';
 import { IdentityInput } from '@concordium/common-sdk';
 import { decryptMobileWalletExport, EncryptedData } from '../src/wallet/crypto';
 import { MobileWalletExport } from '../src/wallet/types';
+import createConcordiumClientV2 from '../src/clientV2';
+import ConcordiumNodeClientV2 from '@concordium/common-sdk/lib/GRPCClient';
 
 export { getModuleBuffer } from '../src/util';
 
@@ -23,6 +25,24 @@ export function getNodeClient(
         credentials.createInsecure(),
         metadata,
         15000
+    );
+}
+
+/**
+ * Creates a client to communicate with a local concordium-node
+ * used for automatic tests.
+ */
+export function getNodeClientV2(
+    address = 'node.testnet.concordium.com',
+    port = 20000
+): ConcordiumNodeClientV2 {
+    const metadata = new Metadata();
+    return createConcordiumClientV2(
+        address,
+        port,
+        credentials.createInsecure(),
+        metadata,
+        { timeout: 15000 }
     );
 }
